+++ src/renderer/components/Notes/Editor.vue
@@ -159,3 +159,9 @@ export default {
   }
 }
 </script>
+<style lang="css" scoped>
+/* >>> -means- that this style should effect child components */
+>>> .CodeMirror {
+  height: 600px;
+}
+</style>